refactor expiration policy per svc for STs/PTs

+++ support/cas-server-support-oauth-core-api/src/main/java/org/apereo/cas/ticket/code/OAuthCodeExpirationPolicy.java
@@ -23,7 +23,7 @@ public class OAuthCodeExpirationPolicy extends MultiTimeUseOrTimeoutExpirationPo
      * @param numberOfUses             the number of uses
      * @param timeToKillInMilliSeconds the time to kill in milli seconds
      */
-    public OAuthCodeExpirationPolicy(@JsonProperty("numberOfUses") final int numberOfUses,
+    public OAuthCodeExpirationPolicy(@JsonProperty("numberOfUses") final long numberOfUses,
                                      @JsonProperty("timeToLive") final long timeToKillInMilliSeconds) {
         super(numberOfUses, timeToKillInMilliSeconds);
     }
